Updated README

+++ lib/constants/constants.dart
@@ -2,4 +2,4 @@ const String kApiHost = 'api.openweathermap.org';
 const String kIconHost = 'www.openweathermap.org';
 const String kLimit = '1';
 const String kUnits = 'metric';
-const double kWarmOrNot = 20.0;
+const int kWarmOrNot = 20;
